+++ doc/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 0.4.2
+
+* Extensions installed with core are searched too
+* Added new algorithms for Typo3 installation and used path
+* Bugfixes
+
 ## Version 0.4.1
 
 * Fixed link to socksipy for python 3